requirements.txt
fastapi
uvicorn[standard]
python-binance
websockets
numpy
pandas
python-dotenv